ADDED: JWT Token with JMS

+++ src/main/java/com/bridgelabz/addressbook/repository/AddressBookRepository.java
@@ -3,5 +3,5 @@ package com.bridgelabz.addressbook.repository;
 import com.bridgelabz.addressbook.model.AddressBookData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AddressBookRepository extends JpaRepository<AddressBookData,Long> {
+public interface AddressBookRepository extends JpaRepository<AddressBookData, Integer> {
 }
